create props and fix tests
